Modify the part of FIXME
In this version, I modify the part of FIXME in the
tst_ref.c. To let the REF version can work, it still
need to build a two D array to store the input. And
change the argument in the old version to fit the new
condition. Finally remove some bug such like free array's
memerory address.

After finishing modifing code about 'FIXME',I use perf to
observe the performance in CPY and REF version. It shows
that the cache-miss rate is close in the two version.